pyproject.toml
[tool.poetry]
name = "warehouse-django"
version = "0.1.0"
description = ""
authors = ["Patrycja Pytka <patrycjapytka01@gmail.com>"]
readme = "README.md"
packages = [{include = "warehouse_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django-crispy-forms = "^2.0"
django-datatables-view = "^1.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
